+++ packages/multichain-api-middleware/src/handlers/types.ts
@@ -1,8 +1,13 @@
+import {
+  Caip25CaveatType,
+  Caip25CaveatValue,
+} from '@metamask/chain-agnostic-permission';
 import type {
-  CaveatSpecificationConstraint,
-  PermissionController,
-  PermissionSpecificationConstraint,
+  GenericPermissionController,
+  Caveat,
 } from '@metamask/permission-controller';
+import type { MultichainRoutingService } from '@metamask/snaps-controllers';
+import type { CaipAccountId } from '@metamask/utils';
 
 /**
  * Multichain API notifications currently supported by/known to the wallet.
@@ -11,11 +16,24 @@ export enum MultichainApiNotifications {
   sessionChanged = 'wallet_sessionChanged',
   walletNotify = 'wallet_notify',
 }
-type AbstractPermissionController = PermissionController<
-  PermissionSpecificationConstraint,
-  CaveatSpecificationConstraint
->;
+
+export type Caip25Caveat = Caveat<typeof Caip25CaveatType, Caip25CaveatValue>;
 
 export type GrantedPermissions = Awaited<
-  ReturnType<AbstractPermissionController['requestPermissions']>
+  ReturnType<GenericPermissionController['requestPermissions']>
 >[0];
+
+export type GetCaveatForOriginHook = {
+  getCaveatForOrigin: (
+    endowmentPermissionName: string,
+    caveatType: string,
+  ) => ReturnType<GenericPermissionController['getCaveat']>;
+};
+
+export type GetNonEvmSupportedMethodsHook = {
+  getNonEvmSupportedMethods: MultichainRoutingService['getSupportedMethods'];
+};
+
+export type SortAccountIdsByLastSelectedHook = {
+  sortAccountIdsByLastSelected: (accounts: CaipAccountId[]) => CaipAccountId[];
+};